drfcustom/requirements.txt
djangorestframework==3.5.4
